refactor for routers
