+++ routers/bearRouter.js
@@ -1,0 +1,7 @@
+const express = require('express');
+
+const bearRouter = express.Router();
+
+const db = require('../data/dbConfig');
+
+module.exports = bearRouter;
